Implicit stack allocation of non-escaping StringBuilders + two GC-trigger bugs
javac lowers every string concatenation to NEW StringBuilder; DUP; <init>;
append-chain; toString. HotSpot's escape analysis scalar-replaces the whole
thing; this is the AOT equivalent: a per-site escape proof in
BytecodeMethod.stackAllocStringBuilders that the builder reference is only
ever the RECEIVER of java.lang.StringBuilder virtual calls (append returns
`this`, so the returned alias is tracked -- popped, chained, or re-stored
into the SAME local), then lowers the NEW onto the @StackAllocate stack-
struct machinery WITHOUT any annotation (TypeInstruction.markImplicitStackAlloc).
The walk is a small CFG traversal with per-index stack-depth states, so the
ternary-in-argument shape (`sb.append(c ? "a" : "b")`) javac loves is
covered; joins with disagreeing depths, backward edges while the ref is
live, arg-passes, field/array stores, returns, throws, try/catch methods,
and anything outside a conservative stack-effect table all bail to the heap
path. When the ctor's @Fused plan has one CONSTANT-length child (the no-arg
builder's char[INITIAL_CAPACITY] -- read from the ctor bytecode, not
duplicated) the buffer is placed in a stack blob too and pre-installed so
the keep-if-null ctor keeps it: such a builder allocates NOTHING on the
heap; only toString's fused String remains. GC safety: the conservative
scan walks the whole native stack, so the struct's heap references (a grown
buffer) are found; the struct itself has heapPosition -1 and is never swept.
Escape hatch: -DCN1_DISABLE_SB_STACK_ALLOC.

Exposed and fixed two REAL pre-existing GC-trigger bugs:
- allocationsSinceLastGC was an int accumulating BYTES; churn workloads move
  multiple GB between cycles, the counter wrapped NEGATIVE, isHighFrequencyGC
  returned false and the GC thread slept its 30s idle wait while dead pages
  ballooned (measured: 44946 live 64KB pages ~ 2.9GB, ONE cycle in a 6s run,
  mutator stalled in the hard-cap backpressure spin). Now long long.
- cn1BibopMaybeGc returned early in nativeAllocationMode, so a workload
  whose every allocation happens inside a native bracket (StringBuilder.
  toString) never fired the 24MB byte trigger. Under CN1_CONSERVATIVE_GC_ROOTS
  native C locals are scanned, so the trigger and safepoint are safe there --
  the early-return is now legacy-only, and the System.gc() call saves/
  restores the caller's bracket instead of clobbering it to FALSE.

Also: the @StackAllocate NEW's static-initializer call is now guarded by the
initialized flag (it was an unconditional call inside every allocating loop).

SbOnly interleaved: ~33% faster than the fused-builder state; GC pages at
steady state 1379 vs 2969. Bench stringBuilding 2.40x -> 1.91x vs warmed
azul-25 (first time under 2x); geomean 1.16x -> 1.14x (loaded machine,
interleaved ratios).

Validation: MapTorture/FusedTest/IbpTest/ThreadChurn/StrCmp/SbTorture
byte-identical (Map/Fused/SbTorture also forced-signal); GcStress 20/20 +
10/10 SIG; MtStress 10/10 plain+SIG; Bench checksums 10/10 bit-identical
plain + ThinLTO.

+++ vm/ByteCodeTranslator/src/cn1_globals.h
@@ -1189,7 +1189,7 @@ extern void cn1BibopNoteMonitorAttached(JAVA_OBJECT obj);
 extern void* cn1MonitorDataGet(JAVA_OBJECT o);
 extern void cn1MonitorDataSet(JAVA_OBJECT o, void* data);
 extern void* cn1MonitorDataRemove(JAVA_OBJECT o);
-extern int allocationsSinceLastGC;
+extern long long allocationsSinceLastGC;
 extern long long totalAllocations;
 
 // LEVER A (perf-tier1, -DCN1_DEATOMIC_BYTES): per-object BiBOP byte accounting.
@@ -1211,13 +1211,13 @@ extern long long totalAllocations;
     JAVA_LONG __bl = (ts)->bibopBytesLocal; \
     if(__bl) { \
         atomic_fetch_add_explicit(&bibopBytesSinceGc, __bl, memory_order_relaxed); \
-        allocationsSinceLastGC += (int)__bl; \
+        allocationsSinceLastGC += __bl; \
         totalAllocations += __bl; \
         (ts)->bibopBytesLocal = 0; } } while(0)
 #else
 #define CN1_BIBOP_ACCOUNT_BYTES(ts, n) do { \
     atomic_fetch_add_explicit(&bibopBytesSinceGc, (n), memory_order_relaxed); \
-    allocationsSinceLastGC += (int)(n); totalAllocations += (n); } while(0)
+    allocationsSinceLastGC += (n); totalAllocations += (n); } while(0)
 #define CN1_BIBOP_FLUSH_BYTES(ts) do {} while(0)
 #endif
